Projet/requirements_projet.txt
dash
numpy
pandas
plotly
geopy
matplotlib
scikit-learn
seaborn
folium
shapely
python-dotenv